refactor: Changed to own files for each component used on home page
