+++ kawayfsl/Components/ToDoCard.jsx
@@ -1,0 +1,11 @@
+export default function ToDoCard() {
+    return (
+      <div className="todo card">
+        <h2 className="todo-title">To-do</h2>
+        <ul>
+          <li>No items to display</li>
+          <li>No items to display</li>
+        </ul>
+      </div>
+    );
+  }
